added discard warning
added alert when you try to discard node
working on collect mode interface

+++ app/src/components/CardsLayer.vue
@@ -94,7 +94,11 @@ export default {
 
     deleteFlag(e) {
       e = this.nodeid
-      this.$store.dispatch('deleteFlag', { e })
+      if (confirm('Confirm discard?')) {
+        this.$store.dispatch('deleteFlag', { e })
+      } else {
+        // nothing happens
+      }
     },
 
     // readFlag(e) {
